Move plans dashboard views to plans app
