+++ meinberlin/apps/dashboard/urls.py
@@ -4,6 +4,7 @@ from meinberlin.apps.bplan.views import BplanProjectCreateView
 from meinberlin.apps.dashboard2.urls import \
     urlpatterns as a4dashboard_urlpatterns
 from meinberlin.apps.extprojects.views import ExternalProjectCreateView
+from meinberlin.apps.plans import views as plans_views
 from meinberlin.apps.projectcontainers.views import ContainerCreateView
 from meinberlin.apps.projectcontainers.views import ContainerListView
 
@@ -22,18 +23,18 @@ urlpatterns = [
         ContainerListView.as_view(),
         name='container-list'),
     url(r'^organisations/(?P<organisation_slug>[-\w_]+)/plans/$',
-        views.DashboardPlanListView.as_view(),
+        plans_views.DashboardPlanListView.as_view(),
         name='plan-list'),
     url(r'^organisations/(?P<organisation_slug>[-\w_]+)/plans/create/$',
-        views.DashboardPlanCreateView.as_view(),
+        plans_views.DashboardPlanCreateView.as_view(),
         name='plan-create'),
     url(r'^organisations/(?P<organisation_slug>[-\w_]+)'
         r'/plans/(?P<slug>[-\w_]+)/$',
-        views.DashboardPlanUpdateView.as_view(),
+        plans_views.DashboardPlanUpdateView.as_view(),
         name='plan-update'),
     url(r'^organisations/(?P<organisation_slug>[-\w_]+)'
         r'/plans/(?P<slug>[-\w_]+)/delete/$',
-        views.DashboardPlanDeleteView.as_view(),
+        plans_views.DashboardPlanDeleteView.as_view(),
         name='plan-delete'),
 
     # Overwrite adhocracy4 core urls with meinBerlin urls
