+++ meinberlin/apps/dashboard/views.py
@@ -1,4 +1,3 @@
-from django.contrib import messages
 from django.contrib.messages.views import SuccessMessageMixin
 from django.core.urlresolvers import reverse
 from django.utils.translation import ugettext_lazy as _
@@ -9,8 +8,6 @@ from meinberlin.apps.dashboard2 import views as a4dashboard_views
 from meinberlin.apps.newsletters.forms import NewsletterForm
 from meinberlin.apps.newsletters.views import NewsletterCreateView
 from meinberlin.apps.organisations.models import Organisation
-from meinberlin.apps.plans.forms import PlanForm
-from meinberlin.apps.plans.models import Plan
 
 from . import forms
 
@@ -61,77 +58,3 @@ class DashboardNewsletterCreateView(a4dashboard_mixins.DashboardBaseMixin,
 
     def get_permission_object(self):
         return self.organisation
-
-
-class DashboardPlanListView(a4dashboard_mixins.DashboardBaseMixin,
-                            generic.ListView):
-    model = Plan
-    template_name = 'meinberlin_dashboard/plan_list.html'
-    permission_required = 'a4projects.add_project'
-    menu_item = 'project'
-
-    def get_permission_object(self):
-        return self.organisation
-
-    def get_queryset(self):
-        return super().get_queryset().filter(organisation=self.organisation)
-
-
-class DashboardPlanCreateView(a4dashboard_mixins.DashboardBaseMixin,
-                              generic.CreateView):
-    model = Plan
-    form_class = PlanForm
-    permission_required = 'meinberlin_plans.add_plan'
-    template_name = 'meinberlin_dashboard/plan_create_form.html'
-    menu_item = 'project'
-
-    def form_valid(self, form):
-        form.instance.creator = self.request.user
-        form.instance.organisation = self.organisation
-        return super().form_valid(form)
-
-    def get_permission_object(self):
-        return self.organisation
-
-    def get_success_url(self):
-        return reverse(
-            'a4dashboard:plan-list',
-            kwargs={'organisation_slug': self.organisation.slug})
-
-
-class DashboardPlanUpdateView(a4dashboard_mixins.DashboardBaseMixin,
-                              generic.UpdateView):
-    model = Plan
-    form_class = PlanForm
-    permission_required = 'meinberlin_plans.add_plan'
-    template_name = 'meinberlin_dashboard/plan_update_form.html'
-    menu_item = 'project'
-
-    def get_permission_object(self):
-        return self.organisation
-
-    def get_success_url(self):
-        return reverse(
-            'a4dashboard:plan-list',
-            kwargs={'organisation_slug': self.organisation.slug})
-
-
-class DashboardPlanDeleteView(a4dashboard_mixins.DashboardBaseMixin,
-                              generic.DeleteView):
-    model = Plan
-    success_message = _('The plan has been deleted')
-    permission_required = 'meinberlin_plans.change_plan'
-    template_name = 'meinberlin_dashboard/plan_confirm_delete.html'
-    menu_item = 'project'
-
-    def delete(self, request, *args, **kwargs):
-        messages.success(self.request, self.success_message)
-        return super().delete(request, *args, **kwargs)
-
-    def get_permission_object(self):
-        return self.organisation
-
-    def get_success_url(self):
-        return reverse(
-            'a4dashboard:plan-list',
-            kwargs={'organisation_slug': self.organisation.slug})
