+++ meinberlin/apps/plans/views.py
@@ -1,9 +1,15 @@
 import json
 
 from django.conf import settings
+from django.contrib import messages
+from django.urls import reverse
+from django.utils.translation import ugettext_lazy as _
 from django.views import generic
 
 from adhocracy4.rules import mixins as rules_mixins
+from meinberlin.apps.dashboard2 import mixins as a4dashboard_mixins
+from meinberlin.apps.plans.forms import PlanForm
+from meinberlin.apps.plans.models import Plan
 
 from . import models
 
@@ -43,3 +49,77 @@ class PlanListView(rules_mixins.PermissionRequiredMixin,
         context['bounds'] = json.dumps(settings.A4_MAP_BOUNDING_BOX)
 
         return context
+
+
+class DashboardPlanListView(a4dashboard_mixins.DashboardBaseMixin,
+                            generic.ListView):
+    model = Plan
+    template_name = 'meinberlin_dashboard/plan_list.html'
+    permission_required = 'a4projects.add_project'
+    menu_item = 'project'
+
+    def get_permission_object(self):
+        return self.organisation
+
+    def get_queryset(self):
+        return super().get_queryset().filter(organisation=self.organisation)
+
+
+class DashboardPlanCreateView(a4dashboard_mixins.DashboardBaseMixin,
+                              generic.CreateView):
+    model = Plan
+    form_class = PlanForm
+    permission_required = 'meinberlin_plans.add_plan'
+    template_name = 'meinberlin_dashboard/plan_create_form.html'
+    menu_item = 'project'
+
+    def form_valid(self, form):
+        form.instance.creator = self.request.user
+        form.instance.organisation = self.organisation
+        return super().form_valid(form)
+
+    def get_permission_object(self):
+        return self.organisation
+
+    def get_success_url(self):
+        return reverse(
+            'a4dashboard:plan-list',
+            kwargs={'organisation_slug': self.organisation.slug})
+
+
+class DashboardPlanUpdateView(a4dashboard_mixins.DashboardBaseMixin,
+                              generic.UpdateView):
+    model = Plan
+    form_class = PlanForm
+    permission_required = 'meinberlin_plans.add_plan'
+    template_name = 'meinberlin_dashboard/plan_update_form.html'
+    menu_item = 'project'
+
+    def get_permission_object(self):
+        return self.organisation
+
+    def get_success_url(self):
+        return reverse(
+            'a4dashboard:plan-list',
+            kwargs={'organisation_slug': self.organisation.slug})
+
+
+class DashboardPlanDeleteView(a4dashboard_mixins.DashboardBaseMixin,
+                              generic.DeleteView):
+    model = Plan
+    success_message = _('The plan has been deleted')
+    permission_required = 'meinberlin_plans.change_plan'
+    template_name = 'meinberlin_dashboard/plan_confirm_delete.html'
+    menu_item = 'project'
+
+    def delete(self, request, *args, **kwargs):
+        messages.success(self.request, self.success_message)
+        return super().delete(request, *args, **kwargs)
+
+    def get_permission_object(self):
+        return self.organisation
+
+    def get_success_url(self):
+        return reverse(
+            'a4dashboard:plan-list',
+            kwargs={'organisation_slug': self.organisation.slug})
